requirements.txt
torch==2.3.0+cu121
torchvision==0.18.0+cu121
Pillow
einops
safetensors
timm
tomesd
accelerate
torchdiffeq
torchsde
xformers
decord
datasets
numpy
scikit-image
opencv-python
omegaconf
SentencePiece
albumentations
diffusers==0.27.0
transformers==4.37.2
beautifulsoup4
ftfy
exllamav2>=0.0.17